pyproject.toml
[project]
name = "retrieval-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kevin Solorio", email = "103829+ksolo@users.noreply.github.com" }
]
requires-python = ">=3.13.7"
dependencies = [
    "datasets>=4.1.1",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-cli>=0.4.2",
    "langsmith>=0.4.29",
    "openevals>=0.1.0",
    "python-dotenv>=1.1.1",
    "weaviate-client>=4.17.0",
]

[project.scripts]
retrieval-demo = "retrieval_demo:main"
ingest-data = "retrieval_demo.pipeline.run_ingestion:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]
